all tasks were updated. TimeInspector. Archive and CassandraDataControl were improved

+++ src/main/java/com/ssd/mvd/gpstabletsservice/task/findFaceFromAssomidin/car_events/Color.java
@@ -7,5 +7,5 @@ import lombok.*;
 @AllArgsConstructor
 public class Color {
     private String name;
-    private Integer confidence;
+    private Double confidence;
 }
