+++ src/main/java/com/ssd/mvd/gpstabletsservice/task/findFaceFromAssomidin/car_events/LicensePlateRegion.java
@@ -6,6 +6,6 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class LicensePlateRegion {
-    private Integer confidence;
+    private Double confidence;
     private String name;
 }
